backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
langchain==0.3.14
langchain-groq==0.2.4
langgraph==0.2.60
httpx==0.28.1
